이벤트 , SMS 모듈 추가 | 와치독 동작시간 업데이트 추가

+++ src/main/java/com/watchDog/project/model/ChkDbTableSpVo.java
@@ -2,10 +2,10 @@ package com.watchDog.project.model;
 
 public class ChkDbTableSpVo {
 	private String tablespace;
-	private float totalspace;
-	private float usespace;
-	private float freespace;
-	private float usepercent;
+	private double totalspace;
+	private double usespace;
+	private double freespace;
+	private double usepercent;
 	@Override
 	public String toString() {
 		return "ChkDbTableSpVo [tablespace=" + tablespace + ", totalspace=" + totalspace + ", usespace=" + usespace
@@ -17,28 +17,28 @@ public class ChkDbTableSpVo {
 	public void setTablespace(String tablespace) {
 		this.tablespace = tablespace;
 	}
-	public float getTotalspace() {
+	public double getTotalspace() {
 		return totalspace;
 	}
-	public void setTotalspace(float totalspace) {
+	public void setTotalspace(double totalspace) {
 		this.totalspace = totalspace;
 	}
-	public float getUsespace() {
+	public double getUsespace() {
 		return usespace;
 	}
-	public void setUsespace(float usespace) {
+	public void setUsespace(double usespace) {
 		this.usespace = usespace;
 	}
-	public float getFreespace() {
+	public double getFreespace() {
 		return freespace;
 	}
-	public void setFreespace(float freespace) {
+	public void setFreespace(double freespace) {
 		this.freespace = freespace;
 	}
-	public float getUsepercent() {
+	public double getUsepercent() {
 		return usepercent;
 	}
-	public void setUsepercent(float usepercent) {
+	public void setUsepercent(double usepercent) {
 		this.usepercent = usepercent;
 	}
